small refactor in the CLI
