+++ cli/index.js
@@ -2,7 +2,7 @@
 
 var pjson = require('./package.json');
 const prog = require('caporal');
-const encrypt = require('./encrypt');
+const encrypt = require('./actions/encrypt');
 const regexGuid = /^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
 
 prog
